moved getUserSyncs to lib pkg
