+++ modules/mgidXBidAdapter.js
@@ -4,7 +4,7 @@ import { BANNER, NATIVE, VIDEO } from '../src/mediaTypes.js';
 import { config } from '../src/config.js';
 import { USERSYNC_DEFAULT_CONFIG } from '../src/userSync.js';
 import { isBidRequestValid, buildRequestsBase, interpretResponse } from '../libraries/teqblazeUtils/bidderUtils.js';
-import { getUserSyncs } from './mgidBidAdapter.js'
+import { getUserSyncs } from '../libraries/mgidUtils/mgidUtils.js' 
 
 const BIDDER_CODE = 'mgidX';
 const GVLID = 358;
